pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "ghost_brain"
description = "Voice-interviewer bot using Pipecat, Twilio, Deepgram, Groq, and OpenAI TTS."
readme = "README.md"
requires-python = "==3.12.*"
license = "MIT"
authors = [{ name = "Ghost Brain" }]
dynamic = ["version"]
dependencies = [
  "pipecat-ai[daily,deepgram,openai,silero,groq]",
  "fastapi",
  "python-multipart",
  "uvicorn[standard]",
  "google-cloud-storage",
  "pydantic-settings",
  "soxr>=0.3.7",
  "ddgs>=9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "ruff>=0.8",
  "pre-commit>=4.0",


  "zensical",
  "mkdocstrings[python]",
  "pyaudio",  # Only for local dev
]

[project.urls]
Documentation = "https://github.com/ghost_brain/ghost_brain#readme"
Repository = "https://github.com/ghost_brain/ghost_brain"

[tool.hatch.build.targets.wheel]
packages = ["src/ghost_brain"]

[tool.hatch.version]
path = "src/ghost_brain/__about__.py"

[tool.hatch.envs.default]
python = "3.12"
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest"

[tool.hatch.envs.dev]
python = "3.12"
features = ["dev"]

[tool.hatch.envs.dev.scripts]
test = "pytest"
lint = "ruff check src tests"
fmt = "ruff format src tests"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["I", "E", "F", "B", "C4", "SIM", "UP", "ARG", "PTH"]
ignore = ["ARG001", "ARG002", "S101"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src/ghost_brain"]
omit = ["*/__about__.py", "*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise AssertionError"]
